+++ src/Domain/Field/Json.php
@@ -4,15 +4,14 @@ declare(strict_types=1);
 
 namespace SylvainDuval\DynamicDbBundle\Domain\Field;
 
+use SylvainDuval\DynamicDbBundle\Domain\Options\FieldOptionsInterface;
+
 final readonly class Json implements FieldInterface
 {
-	/**
-	 * @param ?array<mixed> $default
-	 */
 	public function __construct(
 		public string $name,
 		public bool $nullable = false,
-		public ?array $default = null,
+		public ?FieldOptionsInterface $options = null,
 	) {}
 
 	public function getName(): string