+++ src/Domain/Options/MariaDb/DatabaseOptions.php
@@ -4,6 +4,13 @@ declare(strict_types=1);
 
 namespace SylvainDuval\DynamicDbBundle\Domain\Options\MariaDb;
 
-use SylvainDuval\DynamicDbBundle\Domain\Options\MySql\DatabaseOptions as MySqlDatabaseOptions;
+use SylvainDuval\DynamicDbBundle\Domain\Options\DatabaseOptionsInterface;
 
-final readonly class DatabaseOptions extends MySqlDatabaseOptions {}
+final readonly class DatabaseOptions implements DatabaseOptionsInterface
+{
+	public function __construct(
+		public ?string $charset = null,
+		public ?string $collation = null,
+		public ?string $comment = null,
+	) {}
+}