+++ src/Schema/Postgres/Field/JsonFieldGenerator.php
@@ -7,6 +7,7 @@ namespace SylvainDuval\DynamicDbBundle\Schema\Postgres\Field;
 use InvalidArgumentException;
 use SylvainDuval\DynamicDbBundle\Domain\Field\FieldInterface;
 use SylvainDuval\DynamicDbBundle\Domain\Field\Json;
+use SylvainDuval\DynamicDbBundle\Domain\Options\Postgres\Field\JsonOptions;
 use SylvainDuval\DynamicDbBundle\Schema\Postgres\FieldGeneratorInterface;
 
 /**
@@ -37,11 +38,15 @@ final class JsonFieldGenerator implements FieldGeneratorInterface
 			throw new InvalidArgumentException('Expected Json, found ' . $field::class);
 		}
 
-		if ($field->default === null && $field->nullable) {
+		if ($field->options !== null && !$field->options instanceof JsonOptions) {
+			throw new InvalidArgumentException('Expected JsonOptions, found ' . $field->options::class);
+		}
+
+		if ($field->options?->default === null && $field->nullable) {
 			return ' DEFAULT NULL';
 		}
-		if ($field->default !== null) {
-			return ' DEFAULT \'' . $this->escapeSingleQuote(\json_encode($field->default, JSON_THROW_ON_ERROR)) . '\'';
+		if ($field->options?->default !== null) {
+			return ' DEFAULT \'' . $this->escapeSingleQuote(\json_encode($field->options->default, JSON_THROW_ON_ERROR)) . '\'';
 		}
 
 		return '';