+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2026-02-14
+
+### Added
+
+- Json field options with default value for MariaDb and Postgres
+- Docker environment for integration tests
+
+### Fixed
+
+- Current date for date field in MySQL
+- Remove comment on database for MySQL
+- Table query for MySQL
+
 ## [0.3.0] - 2025-12-06
 
 ### Changed